Moved the integration tests to the tests folder.
